Separate OrderService and ProductService
